+++ src/runtime/base/file/mem_file.h
@@ -31,7 +31,7 @@ public:
   DECLARE_OBJECT_ALLOCATION(MemFile);
 
   MemFile();
-  MemFile(const char *data, int len);
+  MemFile(const char *data, int64 len);
   virtual ~MemFile();
 
   // overriding ResourceData
@@ -39,11 +39,11 @@ public:
 
   virtual bool open(CStrRef filename, CStrRef mode);
   virtual bool close();
-  virtual int readImpl(char *buffer, int length);
+  virtual int64 readImpl(char *buffer, int64 length);
   virtual int getc();
-  virtual int writeImpl(const char *buffer, int length);
-  virtual bool seek(int offset, int whence = SEEK_SET);
-  virtual int tell();
+  virtual int64 writeImpl(const char *buffer, int64 length);
+  virtual bool seek(int64 offset, int whence = SEEK_SET);
+  virtual int64 tell();
   virtual bool eof();
   virtual bool rewind();
   virtual bool flush();
@@ -53,8 +53,8 @@ public:
 protected:
   std::string m_name; // name of the memory file
   char *m_data;       // data of the memory file
-  int m_len;          // length of the memory file
-  int m_cursor;       // m_data's read position
+  int64 m_len;          // length of the memory file
+  int64 m_cursor;       // m_data's read position
   bool m_malloced;    // whether to free m_data on delete
 
   bool closeImpl();
