+++ src/runtime/base/file/zip_file.h
@@ -38,10 +38,10 @@ public:
 
   virtual bool open(CStrRef filename, CStrRef mode);
   virtual bool close();
-  virtual int readImpl(char *buffer, int length);
-  virtual int writeImpl(const char *buffer, int length);
-  virtual bool seek(int offset, int whence = SEEK_SET);
-  virtual int tell();
+  virtual int64 readImpl(char *buffer, int64 length);
+  virtual int64 writeImpl(const char *buffer, int64 length);
+  virtual bool seek(int64 offset, int whence = SEEK_SET);
+  virtual int64 tell();
   virtual bool eof();
   virtual bool rewind();
   virtual bool flush();
